+++ core/markdownValidator.js
