+++ core/memoryOperations.js
@@ -1,11 +1,11 @@
 const fs = require('fs');
 const path = require('path');
-const github = require('../githubClient');
-const tokenStore = require('../tokenStore');
-const memoryConfig = require('../memoryConfig');
-const mdEditor = require('../markdownEditor');
-const fileEditor = require('../markdownFileEditor');
-const validator = require('../markdownValidator');
+const github = require('../utils/githubClient');
+const tokenStore = require('../utils/tokenStore');
+const memoryConfig = require('../utils/memoryConfig');
+const mdEditor = require('./markdownEditor');
+const fileEditor = require('./markdownFileEditor');
+const validator = require('./markdownValidator');
 const {
   ensureDir,
   deepMerge,
