+++ utils/memoryConfig.js
